HHH-7009 Reorg dialiect class names per Jira feedback
Primary dialects are now versioned with the non-versioned
dialect class remaining but deprecated for compatibility
